requirements.txt
mysql-connector-python==8.0.19
PyYAML==5.3.1